DSL: fix parameter handling
In some cases we would output updates in the form var := parameter,
which is a no-no since "parameter" is really an (input or output)
event.
So, we now check whether each transition refers to parameters; if it
does we split it into 2^n transitions (n = no. of referred params)
and add the values of the parameters to their conditions.