Add block-comments /* ... */ to prog syntax
To add comments, we simply override parsec's whitespace parser. We
allow the possibility of a /* followed by any sequence of characters
ended by */. Thus, these block comments may now appear whenever the
original syntax allowed for whitespace.

We also add a spaces() parser at the head of the main parser, to allow
comments (or mere whitespace) at the top of a .prog file.